requirements.txt
Django==1.8.5
django-suit==0.2.15
django-markdown-deux==1.0.5
django-pagedown==0.1.0
